convert PanicController to IntentService so it starts from ACTION_TRIGGER
In order to start a Service from an Intent, it needs to be an IntentService
which also means that the Service runs in its own process.  Intents and
Handlers become the only way to communicate between the app and the Service
then, so the direct method calls have to be replaced.